+++ content/admin/release-notes.md
@@ -4,4 +4,6 @@ layout: release-notes
 versions:
   enterprise-server: '*'
   github-ae: '*'
+topics:
+  - Enterprise
 ---