+++ translations/de-DE/content/admin/release-notes.md
@@ -1,8 +1,10 @@
 ---
 title: Release notes
-intro: The release notes for {{ allVersions[currentVersion].versionTitle }}.
+intro: 'The release notes for {{ allVersions[currentVersion].versionTitle }}.'
 layout: release-notes
 versions:
   enterprise-server: '*'
+topics:
+  - Enterprise
 ---