+++ translations/es-ES/content/admin/release-notes.md
@@ -4,5 +4,7 @@ intro: 'Las notas de lanzamiento para {{ allVersions[currentVersion].versionTitl
 layout: release-notes
 versions:
   enterprise-server: '*'
+topics:
+  - Enterprise
 ---